Assesments/Financial_Portfolio_Manager/requirements.txt
autogen
streamlit
python-dotenv
google-genai
vertexai